+++ PickMeUp/Features/AddressSearch/MVI/AddressSearchView.swift
@@ -268,283 +268,3 @@ struct AddressSearchView: View {
         }
     }
 }
-
-// MARK: - 검색 결과 행 (NavigationLink용)
-struct AddressResultRowContent: View {
-    let location: Location
-
-    var body: some View {
-        HStack(spacing: 16) {
-            // 위치 아이콘
-            ZStack {
-                Circle()
-                    .fill(Color.deepSprout.opacity(0.1))
-                    .frame(width: 44, height: 44)
-
-                Image(systemName: "mappin.circle.fill")
-                    .font(.system(size: 20))
-                    .foregroundColor(.deepSprout)
-            }
-
-            VStack(alignment: .leading, spacing: 6) {
-                // 주요 주소
-                Text(location.address)
-                    .font(.system(size: 16, weight: .medium))
-                    .foregroundColor(.blackSprout)
-                    .multilineTextAlignment(.leading)
-                    .lineLimit(2)
-
-                // 건물명
-                if let name = location.name, !name.isEmpty {
-                    Text(name)
-                        .font(.system(size: 14))
-                        .foregroundColor(.gray60)
-                        .lineLimit(1)
-                }
-
-                // 좌표 정보
-                Text("위도: \(String(format: "%.6f", location.latitude)), 경도: \(String(format: "%.6f", location.longitude))")
-                    .font(.system(size: 12))
-                    .foregroundColor(.gray45)
-                    .lineLimit(1)
-            }
-
-            Spacer()
-
-            Image(systemName: "chevron.right")
-                .font(.system(size: 14))
-                .foregroundColor(.gray45)
-        }
-        .padding(.horizontal, 20)
-        .padding(.vertical, 16)
-    }
-}
-
-// MARK: - 검색 결과 행
-struct AddressResultRow: View {
-    let location: Location
-    let onTap: () -> Void
-
-    var body: some View {
-        Button(action: onTap) {
-            HStack(spacing: 16) {
-                // 위치 아이콘
-                ZStack {
-                    Circle()
-                        .fill(Color.deepSprout.opacity(0.1))
-                        .frame(width: 44, height: 44)
-
-                    Image(systemName: "mappin.circle.fill")
-                        .font(.system(size: 20))
-                        .foregroundColor(.deepSprout)
-                }
-
-                VStack(alignment: .leading, spacing: 6) {
-                    // 주요 주소 (도로명 주소 우선)
-                    Text(location.address)
-                        .font(.system(size: 16, weight: .medium))
-                        .foregroundColor(.blackSprout)
-                        .multilineTextAlignment(.leading)
-                        .lineLimit(2)
-
-                    // 부가 정보 (건물명이 있다면 표시)
-                    if let name = location.name, !name.isEmpty {
-                        Text(name)
-                            .font(.system(size: 14))
-                            .foregroundColor(.gray60)
-                            .lineLimit(1)
-                    }
-
-                    // 좌표 정보 (작은 글씨로)
-                    Text("위도: \(String(format: "%.6f", location.latitude)), 경도: \(String(format: "%.6f", location.longitude))")
-                        .font(.system(size: 12))
-                        .foregroundColor(.gray45)
-                        .lineLimit(1)
-                }
-
-                Spacer()
-
-                Image(systemName: "arrow.up.right")
-                    .font(.system(size: 14))
-                    .foregroundColor(.gray45)
-            }
-            .padding(.horizontal, 20)
-            .padding(.vertical, 16)
-        }
-        .buttonStyle(PlainButtonStyle())
-    }
-}
-
-// MARK: - 에러 메시지 뷰
-struct ErrorMessageView: View {
-    let message: String
-    let onDismiss: () -> Void
-
-    var body: some View {
-        HStack(spacing: 12) {
-            Image(systemName: "exclamationmark.triangle.fill")
-                .font(.system(size: 16))
-                .foregroundColor(.brightForsythia)
-
-            Text(message)
-                .font(.system(size: 14))
-                .foregroundColor(.blackSprout)
-                .multilineTextAlignment(.leading)
-
-            Spacer()
-
-            Button(action: onDismiss) {
-                Image(systemName: "xmark")
-                    .font(.system(size: 12))
-                    .foregroundColor(.gray45)
-            }
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 12)
-        .background(
-            RoundedRectangle(cornerRadius: 8)
-                .fill(Color.brightForsythia.opacity(0.1))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 8)
-                        .stroke(Color.brightForsythia.opacity(0.3), lineWidth: 1)
-                )
-        )
-    }
-}
-
-// MARK: - 주소 검색 상태 관리
-struct AddressSearchState {
-    var searchResults: [Location] = []
-    var isLoading: Bool = false
-    var errorMessage: String? = nil
-    var hasSearched: Bool = false
-}
-
-enum AddressSearchAction {
-    enum Intent {
-        case searchAddress(String)
-        case clearResults
-        case clearError
-    }
-
-    enum Result {
-        case searchStarted
-        case searchSucceeded([Location])
-        case searchFailed(String)
-        case resultsCleared
-        case errorCleared
-    }
-}
-
-struct AddressSearchEffect {
-    func handle(_ intent: AddressSearchAction.Intent, store: AddressSearchStore) {
-        switch intent {
-        case .searchAddress(let query):
-            Task {
-                await searchAddressFromNaverAPI(query: query, store: store)
-            }
-        case .clearResults:
-            Task { @MainActor in
-                store.send(.resultsCleared)
-            }
-        case .clearError:
-            Task { @MainActor in
-                store.send(.errorCleared)
-            }
-        }
-    }
-
-    // 네이버 지오코딩 API 호출
-    private func searchAddressFromNaverAPI(query: String, store: AddressSearchStore) async {
-        await MainActor.run {
-            store.send(.searchStarted)
-        }
-
-        do {
-            let results = try await NaverGeocodingService.shared.searchAddress(query: query)
-
-            await MainActor.run {
-                if results.isEmpty {
-                    store.send(.searchFailed("검색 결과를 찾을 수 없습니다."))
-                } else {
-                    store.send(.searchSucceeded(results))
-                }
-            }
-        } catch let error as NaverGeocodingError {
-            await MainActor.run {
-                store.send(.searchFailed(error.localizedDescription))
-            }
-        } catch {
-            await MainActor.run {
-                store.send(.searchFailed("네트워크 오류가 발생했습니다. 다시 시도해주세요."))
-            }
-        }
-    }
-}
-
-struct AddressSearchReducer {
-    func reduce(state: inout AddressSearchState, intent: AddressSearchAction.Intent) {
-        switch intent {
-        case .searchAddress:
-            state.isLoading = true
-            state.errorMessage = nil
-
-        case .clearResults:
-            state.searchResults = []
-            state.hasSearched = false
-            state.errorMessage = nil
-
-        case .clearError:
-            state.errorMessage = nil
-        }
-    }
-
-    func reduce(state: inout AddressSearchState, result: AddressSearchAction.Result) {
-        switch result {
-        case .searchStarted:
-            state.isLoading = true
-            state.errorMessage = nil
-
-        case .searchSucceeded(let results):
-            state.searchResults = results
-            state.isLoading = false
-            state.hasSearched = true
-
-        case .searchFailed(let error):
-            state.errorMessage = error
-            state.isLoading = false
-            state.hasSearched = true
-
-        case .resultsCleared:
-            state.searchResults = []
-            state.hasSearched = false
-            state.errorMessage = nil
-
-        case .errorCleared:
-            state.errorMessage = nil
-        }
-    }
-}
-
-final class AddressSearchStore: ObservableObject {
-    @Published private(set) var state: AddressSearchState
-    private let effect: AddressSearchEffect
-    private let reducer: AddressSearchReducer
-
-    init(state: AddressSearchState, effect: AddressSearchEffect, reducer: AddressSearchReducer) {
-        self.state = state
-        self.effect = effect
-        self.reducer = reducer
-    }
-
-    @MainActor
-    func send(_ action: AddressSearchAction.Intent) {
-        reducer.reduce(state: &state, intent: action)
-        effect.handle(action, store: self)
-    }
-
-    @MainActor
-    func send(_ result: AddressSearchAction.Result) {
-        reducer.reduce(state: &state, result: result)
-    }
-}
