+++ PickMeUp/Features/AddressSearch/NaverGeocodingService.swift
@@ -9,37 +9,6 @@ import SwiftUI
 import NMapsMap
 import CoreLocation
 
-// MARK: - 네이버 지오코딩 응답 모델
-struct NaverGeocodingResponse: Codable {
-    let status: String
-    let meta: Meta
-    let addresses: [Address]
-    let errorMessage: String?
-
-    struct Meta: Codable {
-        let totalCount: Int
-        let page: Int
-        let count: Int
-    }
-
-    struct Address: Codable {
-        let roadAddress: String?
-        let jibunAddress: String?
-        let englishAddress: String?
-        let addressElements: [AddressElement]
-        let x: String // 경도
-        let y: String // 위도
-        let distance: Double?
-
-        struct AddressElement: Codable {
-            let types: [String]
-            let longName: String
-            let shortName: String
-            let code: String?
-        }
-    }
-}
-
 // MARK: - 네이버 지오코딩 서비스
 final class NaverGeocodingService {
     static let shared = NaverGeocodingService()
@@ -244,114 +213,5 @@ final class NaverGeocodingService {
         }
 
         return addressComponents.joined(separator: " ")
-    }
-}
-
-// MARK: - 역지오코딩 응답 모델
-struct NaverReverseGeocodingResponse: Codable {
-    let status: Status
-    let results: [Result]
-
-    struct Status: Codable {
-        let code: Int
-        let name: String
-        let message: String
-    }
-
-    struct Result: Codable {
-        let name: String
-        let code: Code
-        let region: Region?
-        let land: Land?
-
-        struct Code: Codable {
-            let id: String
-            let type: String
-            let mappingId: String
-        }
-
-        struct Region: Codable {
-            let area0: Area?
-            let area1: Area?
-            let area2: Area?
-            let area3: Area?
-            let area4: Area?
-
-            struct Area: Codable {
-                let name: String
-                let coords: Coords
-                let alias: String?
-
-                struct Coords: Codable {
-                    let center: Center
-
-                    struct Center: Codable {
-                        let crs: String
-                        let x: Double
-                        let y: Double
-                    }
-                }
-            }
-        }
-
-        struct Land: Codable {
-            let type: String
-            let number1: String?
-            let number2: String?
-            let addition0: Addition?
-            let addition1: Addition?
-            let addition2: Addition?
-            let addition3: Addition?
-            let addition4: Addition?
-            let coords: Coords?
-
-            struct Addition: Codable {
-                let type: String
-                let value: String
-            }
-
-            struct Coords: Codable {
-                let center: Center
-
-                struct Center: Codable {
-                    let crs: String
-                    let x: Double
-                    let y: Double
-                }
-            }
-        }
-    }
-}
-
-// MARK: - 에러 정의
-enum NaverGeocodingError: LocalizedError {
-    case invalidQuery
-    case invalidURL
-    case invalidResponse
-    case unauthorized
-    case rateLimitExceeded
-    case serverError(Int)
-    case apiError(String)
-    case networkError(Error)
-
-    var errorDescription: String? {
-        switch self {
-        case .invalidQuery:
-            return "잘못된 검색어입니다."
-        case .invalidURL:
-            return "잘못된 URL입니다."
-        case .invalidResponse:
-            return "서버 응답이 올바르지 않습니다."
-        case .unauthorized:
-            return "API 인증에 실패했습니다. API 키를 확인해주세요."
-        case .rateLimitExceeded:
-            return "API 호출 한도를 초과했습니다. 잠시 후 다시 시도해주세요."
-        case .serverError(let code):
-            return "서버 오류가 발생했습니다. (코드: \(code))"
-        case .apiError(let message):
-            return "API 오류: \(message)"
-        case .networkError(let error):
-            return "네트워크 오류: \(error.localizedDescription)"
-        }
     }
 }
