+++ PickMeUp/Features/LocationSelection/Presentation/MVI/LocationSelectionView.swift
@@ -47,64 +47,6 @@ struct LocationSelectionView: View {
     }
 }
 
-// MARK: - Location Row with Tap Action
-private struct LocationRow: View {
-    let currentLocation: Location
-    let onLocationTap: () -> Void
-
-    var body: some View {
-        Button(action: onLocationTap) {
-            HStack(spacing: 8) {
-                Image("annotation")
-                    .scaledToFit()
-                    .frame(width: 24, height: 24)
-
-                Text(currentLocation.displayName)  // displayName 사용
-                    .font(.pretendardBody1)
-                    .foregroundColor(.gray90)
-
-                Image(systemName: "arrowtriangle.down.fill")
-                    .font(.system(size: 8))
-                    .foregroundColor(.gray90)
-
-                Spacer()
-            }
-        }
-        .buttonStyle(PlainButtonStyle())
-    }
-}
-
-// MARK: - Search Bar
-struct CustomSearchBar: View {
-    @Binding var text: String
-
-    var body: some View {
-        HStack(spacing: 8) {
-            Image(systemName: "magnifyingglass")
-                .foregroundColor(.blackSprout)
-
-            if text.isEmpty {
-                Text("검색어를 입력해주세요.")
-                    .foregroundColor(.gray60)
-                    .font(.pretendardBody2)
-            } else {
-                TextField("", text: $text)
-                    .font(.pretendardBody2)
-                    .foregroundColor(.gray100)
-            }
-
-            Spacer()
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 10)
-        .background(
-            RoundedRectangle(cornerRadius: 20)
-                .stroke(Color.deepSprout, lineWidth: 1)
-                .background(RoundedRectangle(cornerRadius: 20).fill(Color.white))
-        )
-    }
-}
-
 #Preview {
     LocationSelectionView()
 }
