+++ PickMeUp/Features/NaverMap/Presentation/MVI/MapConstants.swift
@@ -1,0 +1,18 @@
+//
+//  MapConstants.swift
+//  PickMeUp
+//
+//  Created by 김태형 on 7/2/25.
+//
+
+import CoreLocation
+import Foundation
+
+enum MapConstants {
+    static let defaultZoom: Double = 16
+    static let seoulCoordinate = CLLocationCoordinate2D(latitude: 37.5666805, longitude: 126.9784147)
+    static let buttonHeight: CGFloat = 50
+    static let horizontalPadding: CGFloat = 20
+    static let bottomPadding: CGFloat = 20
+    static let mapButtonSpacing: CGFloat = 20
+}
