+++ PickMeUp/Features/NaverMap/Presentation/MVI/MapMarkerType.swift
@@ -1,0 +1,14 @@
+//
+//  MapMarkerType.swift
+//  PickMeUp
+//
+//  Created by 김태형 on 7/2/25.
+//
+
+import Foundation
+
+enum MapMarkerType {
+    case currentLocation
+    case centerPin
+    case selectedLocation
+}
